pyproject.toml
[project]
name = "langchain-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
]
